requirements-rl.txt
accelerate==1.4.0
apex==0.9.10dev
attrdict==2.0.1
datasets==3.5.0
decord==0.6.0
deepspeed==0.15.4
einops==0.8.1
fastchat==0.1.0
flash_attn==2.7.4.post1
huggingface_hub==0.30.2
imageio==2.37.0
liger_kernel==0.5.3
numpy==2.3.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
orjson==3.10.18
packaging==25.0
peft==0.10.0
petrel_client==0.0.1
Pillow==11.3.0
Requests==2.32.4
sentencepiece==0.2.0
sympy==1.13.1
timm==1.0.16
tokenizers==0.21.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.50.0
trl==0.16.0
vllm==0.7.2
wandb==0.21.0